requirements.txt
ultralytics
contourpy==1.1.0
openpyxl==3.1.2
cycler==0.11.0
fonttools==4.42.1
kiwisolver==1.4.5
matplotlib==3.7.2
numpy==1.25.2
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
